refactor(payments): fix architecture test compliance
- Rename payment-provider.factory.ts → payment-provider-factory.adapter.ts
- Rename PaymentProviderFactory class → PaymentProviderFactoryAdapter
- Rename infrastructure event handlers from .handler.ts → .listener.ts
- Extract helper interfaces from .port.ts files into application/types/
- Convert exported interfaces in .command.ts/.entity.ts to type aliases
- All 34 architecture fitness tests now pass

Issue #18
