+++ backend/.env.example
@@ -58,10 +58,47 @@ SMTP_PASS=tickr123
 SMTP_FROM=noreply@tickr.local
 SMTP_SECURE=false
 
-# Stripe Payment
+# ============================================
+# Payment Gateways
+# ============================================
+
+# Stripe (International payments - EUR/USD)
 STRIPE_SECRET_KEY=sk_test_your_stripe_secret_key
 STRIPE_PUBLIC_KEY=pk_test_your_stripe_public_key
 STRIPE_WEBHOOK_SECRET=whsec_your_webhook_secret
+
+# Konnect (Primary TN gateway - TND)
+KONNECT_API_URL=https://api.preprod.konnect.network/api/v2
+KONNECT_API_KEY=your_konnect_api_key
+KONNECT_WALLET_ID=your_konnect_wallet_id
+KONNECT_WEBHOOK_SECRET=your_konnect_webhook_secret
+KONNECT_WEBHOOK_URL=https://api.tick-r.tn/api/payments/webhooks/konnect
+
+# Paymee (TN fallback gateway - TND)
+PAYMEE_API_URL=https://sandbox.paymee.tn/api/v2
+PAYMEE_API_KEY=your_paymee_api_key
+PAYMEE_WEBHOOK_URL=https://api.tick-r.tn/api/payments/webhooks/paymee
+
+# Payment URLs (redirect after payment)
+PAYMENT_SUCCESS_URL=https://tick-r.tn/payments/success
+PAYMENT_FAIL_URL=https://tick-r.tn/payments/failure
+
+# ============================================
+# Payment Business Configuration
+# ============================================
+
+# Order settings
+ORDER_EXPIRATION_MINUTES=15
+
+# Fraud detection
+FRAUD_MAX_ORDERS_PER_HOUR=5
+FRAUD_MAX_TICKETS_PER_EVENT=10
+FRAUD_HIGH_VALUE_THRESHOLD_TND=5000
+FRAUD_HIGH_VALUE_THRESHOLD_EUR=2000
+FRAUD_HIGH_VALUE_THRESHOLD_USD=2000
+
+# Scheduler
+SCHEDULER_ORDERS_ENABLED=true
 
 # Rate Limiting
 THROTTLE_TTL=60
@@ -75,8 +112,8 @@ SESSION_MAX_AGE=86400000
 SWAGGER_ENABLE=true
 SWAGGER_PATH=api/docs
 
-# Business Configuration
-PLATFORM_COMMISSION_RATE=0.06
+# Business Configuration (Commission)
+PLATFORM_COMMISSION_RATE=0.04
 PLATFORM_COMMISSION_MIN=0.00
 PLATFORM_COMMISSION_MAX=0.20
