+++ backend/src/modules/payments/application/commands/expire-orders/expire-orders.command.ts
@@ -1,8 +1,8 @@
 import { BaseCommand } from '@shared/application/interfaces/command.interface';
 
-export interface ExpireOrdersResult {
+export type ExpireOrdersResult = {
   readonly expiredCount: number;
-}
+};
 
 export class ExpireOrdersCommand extends BaseCommand {
   constructor() {
