+++ backend/src/modules/payments/application/types/ticket-reservation.types.ts
@@ -1,0 +1,10 @@
+/**
+ * Ticket Reservation Types
+ *
+ * DTOs used by the Ticket Reservation port.
+ */
+
+export interface TicketReservationResult {
+  ticketIds: string[];
+  reservedUntil: Date;
+}
